// src/types/types.ts
export type SvgType =
  | "ChangeMarking"
  | "CloseArrow"
  | "Enters"
  | "EnterLogo"
  | "PaginationArrow"
  | "Pictures"
  | "User"
  | "Users"
  | "More"
  | "Search"
  | "Arrow"
  | "Close"
  | "Eye"
  | "ChangeLog"
  | "SmallLogo"
  | "BigLogo"
  | "Pencil"
  | "Checkmark"
  | "Cube"
  | "FullScreenClosed"
  | "FullScreenOpened"
  | "Warning";